profile page dodan, izmjenjen settings page
